Feat: acquireStickerFromPuzzleResult 수행 위치 변경
