+++ src/main/java/ticketaka/mtvs3_final_backend/member/command/application/controller/MemberCommandController.java
@@ -3,35 +3,17 @@ package ticketaka.mtvs3_final_backend.member.command.application.controller;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import ticketaka.mtvs3_final_backend._core.utils.ApiUtils;
-import ticketaka.mtvs3_final_backend.member.command.application.dto.MemberCommandRequestDTO;
-import ticketaka.mtvs3_final_backend.member.command.application.dto.MemberCommandResponseDTO;
 import ticketaka.mtvs3_final_backend.member.command.application.service.MemberCommandService;
 
-import static ticketaka.mtvs3_final_backend._core.utils.SecurityUtils.getCurrentMemberId;
 
 @Slf4j
 @RequiredArgsConstructor
 @RestController
-@RequestMapping("/api/member")
+@RequestMapping("/api/concert")
 @Tag(name = "07_MemberCommandController")
 public class MemberCommandController {
 
     private final MemberCommandService memberCommandService;
-
-    /*
-        Puzzle 결과 Sticker 획득
-     */
-    @PostMapping("/puzzle/result/sticker")
-    public ResponseEntity<?> acquireStickerFromPuzzleResult(@RequestBody MemberCommandRequestDTO.acquireStickerFromPuzzleResultDTO requestDTO) {
-
-        MemberCommandResponseDTO.acquireStickerFromPuzzleResultDTO responseDTO = memberCommandService.acquireStickerFromPuzzleResult(getCurrentMemberId(), requestDTO);
-
-        return ResponseEntity.ok().body(ApiUtils.success(responseDTO));
-    }
 }
