+++ src/main/java/ticketaka/mtvs3_final_backend/member/command/application/dto/MemberCommandRequestDTO.java
@@ -1,9 +1,4 @@
 package ticketaka.mtvs3_final_backend.member.command.application.dto;
 
 public class MemberCommandRequestDTO {
-
-    public record acquireStickerFromPuzzleResultDTO(
-            int rank
-    ) {
-    }
 }
