+++ src/main/java/ticketaka/mtvs3_final_backend/member/command/application/dto/MemberCommandResponseDTO.java
@@ -1,12 +1,4 @@
 package ticketaka.mtvs3_final_backend.member.command.application.dto;
 
 public class MemberCommandResponseDTO {
-
-    public record acquireStickerFromPuzzleResultDTO(
-            int stickerId,
-            String stickerName,
-            String stickerScript,
-            String stickerRarity
-    ) {
-    }
 }
