+++ src/main/java/ticketaka/mtvs3_final_backend/member/command/application/service/MemberCommandService.java
@@ -1,0 +1,13 @@
+package ticketaka.mtvs3_final_backend.member.command.application.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Slf4j
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+@Service
+public class MemberCommandService {
+}
