+++ src/main/java/ticketaka/mtvs3_final_backend/sticker/query/service/StickerQueryService.java
@@ -9,8 +9,8 @@ import ticketaka.mtvs3_final_backend._core.error.exception.Exception401;
 import ticketaka.mtvs3_final_backend.member.command.domain.model.Member;
 import ticketaka.mtvs3_final_backend.member.query.repository.MemberQueryRepository;
 import ticketaka.mtvs3_final_backend.sticker.command.domain.model.Sticker;
+import ticketaka.mtvs3_final_backend.sticker.command.domain.model.StickerRarity;
 import ticketaka.mtvs3_final_backend.sticker.command.domain.model.StickerType;
-import ticketaka.mtvs3_final_backend.sticker.member.command.domain.model.MemberSticker;
 import ticketaka.mtvs3_final_backend.sticker.query.repository.StickerQueryRepository;
 import ticketaka.mtvs3_final_backend.ticketing.concert.command.domain.model.Concert;
 import ticketaka.mtvs3_final_backend.ticketing.concert.command.domain.model.ConcertStatus;
@@ -18,7 +18,6 @@ import ticketaka.mtvs3_final_backend.ticketing.concert.query.repositroy.ConcertQ
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 @Slf4j
 @Transactional(readOnly = true)
@@ -45,6 +44,12 @@ public class StickerQueryService {
         stickerList.addAll(getMemberStickerList(memberId));
 
         return stickerList;
+    }
+
+    // Sticker 할당
+    public Sticker getPuzzleResult(Long memberId, StickerRarity stickerRarity) {
+
+        return null;
     }
 
     // Member 조회
